Separated vpn apis from manager
